New config system
git-svn-id: svn+ssh://svn.code.sf.net/p/virtualgl/code/trunk@1335 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b

+++ rr/rrdisplayclient.cpp
@@ -80,7 +80,8 @@ rrdisplayclient::rrdisplayclient(char *displayname) : _np(fconfig.np),
 {
 	memset(&_v, 0, sizeof(rrversion));
 	if(fconfig.verbose)
-		rrout.println("[VGL] Using %d / %d CPU's for compression", fconfig.np, numprocs());
+		rrout.println("[VGL] Using %d / %d CPU's for compression",
+			(int)fconfig.np, numprocs());
 	char *servername=NULL;
 	try
 	{
